deparse/core: have ALTER TABLE return address of affected object
Currently, most if not all CREATE commands return the ObjectAddress of
the created object.  This is useful for event triggers to try to figure
out exactly what happened.  This patch extends this idea a bit further
to cover ALTER TABLE as well: an auxiliary ObjectAddress is also
returned for several subcommands of ALTER TABLE.  This address can later
be used to determine, for instance, which constraint was added by ALTER
TABLE ADD CONSTRAINT, or which parent got dropped from the inherit-from
parents list by NO INHERIT.  This makes it possible to decode with
precision exactly what happened during execution of any ALTER TABLE
command.

There is no user-visible change here; this patch makes the info
available for later.